Refactor indicator
