+++ client/coral-admin/src/containers/Header.js
@@ -1,22 +1,18 @@
 import { gql } from 'react-apollo';
 import withQuery from 'coral-framework/hocs/withQuery';
 import Header from '../components/Header';
+import CommunityIndicator from '../routes/Community/containers/Indicator';
+import ModerationIndicator from '../routes/Moderation/containers/Indicator';
+import { getDefinitionName } from 'coral-framework/utils';
 
 export default withQuery(
   gql`
     query TalkAdmin_Header {
-      __typename
-      premodCount: commentCount(query: { statuses: [PREMOD] })
-      reportedCount: commentCount(
-        query: { statuses: [NONE, PREMOD, SYSTEM_WITHHELD], action_type: FLAG }
-      )
-      flaggedUsernamesCount: userCount(
-        query: {
-          action_type: FLAG
-          state: { status: { username: [SET, CHANGED] } }
-        }
-      )
+      ...${getDefinitionName(ModerationIndicator.fragments.root)}
+      ...${getDefinitionName(CommunityIndicator.fragments.root)}
     }
+    ${ModerationIndicator.fragments.root}
+    ${CommunityIndicator.fragments.root}
   `,
   {
     options: {
